api,params: change data size to size_t
Data type parameters are used in 2 node types: AnimKeyFrameBuffer and
all the Buffer* nodes.

AnimKeyFrameBuffer is using animkeyframe_opts to store the parameter,
which is shared amongst animation routines. AnimatedBuffer reads the
data field and transfers it to a buffer_info, so this struct needed an
update of its data_size as well.

Buffer* use a local buffer_opts, and similary to AnimatedBuffer
transfers the data field into a buffer_info.

+++ libnopegl/src/node_animatedbuffer.c
@@ -111,7 +111,7 @@ static int animatedbuffer_init(struct ngl_node *node)
     for (int i = 0; i < o->nb_animkf; i++) {
         const struct animkeyframe_opts *kf = o->animkf[i]->opts;
         const size_t data_count = kf->data_size / layout->stride;
-        const int data_pad   = (int)(kf->data_size % layout->stride);
+        const size_t data_pad   = kf->data_size % layout->stride;
 
         if (layout->count && layout->count != data_count) {
             static const char *types[] = {"float", "vec2", "vec3", "vec4"};
@@ -122,7 +122,7 @@ static int animatedbuffer_init(struct ngl_node *node)
         }
 
         if (data_pad)
-            LOG(WARNING, "the data buffer has %d trailing bytes", data_pad);
+            LOG(WARNING, "the data buffer has %zd trailing bytes", data_pad);
 
         layout->count = data_count;
     }
